Create environment file.
New environment YAML file that contains the description of a first-pass conda environment for the assembly pipeline.
All tools that are in the draft pipeline (i.e samtools, trimmomatic, varscan) are available through conda, however more work may need to be done to make sure that the correct versions of each are being used. For now, defaults will work.