Update reference genomes and ignore index files.
Reference genome chromosome names needed to be updated to fix bugs that were coming up with Picard. Because the binary index files were causing a lot of unnecessary git interfacing they are now ignored as derived files. In the future the generation of index files should be included in the basic pipeline.