Update config format slightly.
The configuration JSON now lives in a separate directory to prepare for future, different configs.
Each reference virus now contains a dictionary of name and total length, which will be used in downstream summary statistic generation.
New references are added to the config and Snakefile is updated to reflect the new filepath.