Add rule `fasta headers`
Edit headers to include the NWGC sample ID.

Uses `seqkit` to replace NWGC ID with SFS barcode.
This requires a tab-delimited key-value file where
key is the NWGC ID and value is the matching
SFS barcode. 

The path to the key-value file must be added
to the config file under the "barcode_match" key.
Add wildcard to temp file