Edit script to include check on paired reads
Rename `flu_sample_ref_pair` to 'setup_config_file`
to more accurately describe what the script does.

This script updates the Snakemake config file by 
filling in the `sample_reference_pairs` and the
`ignored_samples`.

Samples are ignored if they cannot be found in the
sample/target map fed into the script or if the 
sample's R1 and R2 reads differ by over 20%. 

The new json file added to the references directory
is the map used to match target names from the
presence/absence tests to the actual reference
names used in the assembly pipeline. Currently
it only contains Flu targets/references.