rule pileup: Add min base quality
The default base quality (-Q) for `samtools mpileup`
is 15. We want it to match the minimum base quality
used in `varscan mpileup2snp` so that both will 
have the same coverage depth per base. 

The additional `-a` flag makes sure that all
positions are included in the pileup, including
positions with 0 coverage depth.