rule create_bed_file: Generate bed file from pileup
This rule replaces the `rule coverage` and `rule low_coverage`.
This was done because `bedtools genomecov` calculates coverage 
depth without considering base quality. 
This causes the coverage depth generated
to be much higher than the coverage depth considered
within `varscan mpileup2snp` and therefore the 
consensus genome does not get masked appropriately. 

The custom python script creates a bed file based on coverage
and variant counts in the pileup file. If a position has coverage
below the `min_cov`set in the config or if it is disproportionally
supported by one strand (>90%) then it is included in the bed file.
This ensures that the SNPs that fail the `--strand-filter` in 
`varscan mpileup2snp` is included in the bed file and
are masked appropriately in the consensus genome.