Add steps to upload consensus genomes to ID3C
Add five steps to the Snakefile:
1. an intermediate step to create a Json file containing metadata like
URLs
2. an intermediate step to convert the masked consensus fasta file to
Json format
3. an intermediate step to convert the bowtie2 and BAMstats summary
statistics to Json format
4. an intermediate checkpoint to combine the Json files from steps 1, 2,
and 3 to create the final ID3C payload.
5. a final step to upload the newly created payload from step 4 to ID3C
via a HTTP POST request to a `/receiving/consensus-genome` endpoint. The
connection details including specific URL and username/password are
stored in the environment.

If step 5 receives a non-successful HTTP response, then the
response from the URL is stored in a log under the given wildcard
directory/file path. The job does not appear as failed in the Snakemake
DAG.

Co-authored-by: Jover Lee <joverlee521@gmail.com>