Create one sample Snakefile
It may be useful to keep separate Snakefiles for batch
assembly jobs and single sample assembly jobs, so create
a new Snakefile specifically for one sample jobs.

The major change from the original snakefile is the
deletion of the helper functions that were used
to generate all the sample/reference pairings.
Instead, the sample, references, and FASTA files are
directly provided to the assembly job through the
config file that was generated for single samples.

The intermediate files are now stored under directory
with the current date for easier digging through data
at a later date.