Change checkpoint to check mapped reads
Change checkpoint in assembly pipeline to check the number
of mapped reads in the bowtie2 summary instead of the
overall alignment rate. This can avoid situations where samples
with too few reads generate an erroneous consensus genome
and where samples with a lot of reads have a low overall
alignment rate but still have enough mapped reads to 
generate a consensus genome.