Create Slack alerts for failed assembly uploads
If a POST request to the ID3C consensus genome web API endpoint returns
a non-successful HTTP response during an assembly job, create an alert
in the #id3c-alerts Slack channel notifying the user of their failed
job.

This feature is necessary because the Snakemake DAG does not show failed
POST requests as failed jobs in the DAG. Therefore, the current solution
is to bring attention to failed uploads via Slack.

The Slack webhook is stored in the environment.

Co-authored-by: Jover Lee <joverlee521@gmail.com>