setup_config_file: search or read IDs more specifically
Read IDs in FASTQ files start with the `@` symbol, but this can also
be true for the quality values line since the `@` symbol is also a
quality value character.

To differentiate between the read IDs and the quality value lines,
also search for a space in the line since that is not a quality value
character.