Add new checkpoint to verify Flu B lineage based on percent identity
The checkpoint `percent_identity` creates a file that contains the
percent identity of a sample consensus genome compared to the specified
`segment` of the reference genome. The percent identity is compared to
the minimum requirement and the consensus genome does not get uploaded
to ID3C if it is less than the minimum.

We set the references, segment and minimum percent identity in the
configfile. For Vic and Yam references, we calculate the percent
identity for the HA segment. @trvrb recommends to set the minimum
percent identity at 95% since Vic and Yam are ~10% divergent.