Use bcftools `--mask` option to mask consensus
With the change to using `mpileup2cns`, the consensus genome may contain
indels that shift the bases such that the BEDfile generated for masking
no longer aligns correctly. This causes the `bedtools maskfasta` to mask
incorrectly.

This commit changes the pipeline to mask the consensus using the
`--mask` option for the `bcftools consensus` command. This command masks
the sites first before applying variants to the reference. If a site is
already masked, then the variant is skipped (based on the source code
at https://github.com/samtools/bcftools/blob/develop/consensus.c)