Add the `-B` option for `samtools mpileup`
Use the `-B` option to disable BAQ computation in the `samtools mpileup`
command. According to [VarScan documentation](http://varscan.sourceforge.net/germline-calling.html#germline-input),
the BAQ computation is too stringent and reduces the base quality scores
too aggressively.

Using this option allows more deletions to come through in the
generated pileup file for B.1.1.7 and B.1.351 SARS-CoV-2 variants.
This increases the coverage for the deletions so that they can be
included in the final consensus genome.