submissions: updates for new WA DOH template
Update with new column names and new date format.
Add additional step for getting the PANGO lineages since these are
now required for WA DOH so they do not need to rely on GISAID pulls.

Note that PII fields are hard-coded to "N/A" since we do not ingest
PII into Samplify or ID3C.