submissions: Add BioSample and BioProject
Create TSVs for BioSample submissions and add returned BioSample
accessions to GenBank metadata. This allows us to add originating lab
for samples in the "collected_by" field of the BioSample record.
Also includes the SFS BioProject accession in NCBI submissions so that
everything can be linked the shared BioProject.

This does require submissions to BioSample first and wait for them to be
processed. Once the BioSample records have been assigned
accessions, submit sequences to GenBank with the linked BioSample
accession.