create_submissions: create separate SCH report
If SCH samples are in a sequencing batch, create a separate sequencing
results Excel report for them. This will get emailed to SCH so that they
can fill in the original lab accession ids for the samples. SFS cannot
have access to these ids since SCH considers them to be PII.

All other sequencing results will be in another Excel file to be
uploaded to WA DOH SFT as usual.