Add check for empty dataframe before creating final submission outputs
If there are no new submissions (with status=`submitted`), the create_biosample_submission
function throws an error. This along with `create_gisaid_submission`, and `create_genbank_submission`
don't seem necessary if there are no new submissions, so exiting preemptively with message indicating
that with a summary of the status counts.