Add QC of external metadata Excel file
Standardize and validate barcode and collection date values in the external metadata
Excel file. The Excel file contains 2 sheets which are loaded to dataframes, then written back
to Excel after minor transformations and passing all validation checks.

Barcodes are validated as 8-digit hexidecimal values. Collection dates are updated to remove
"na" values and confirm no dates prior to 2020-02-01 and that not all values are identical.