Run `export_id3c_metadata` script as part of file prep
Optionally run `export_id3c_metadata` bash script while running the file
prep script to avoid running as a separate command.

This temporarily redirects stdout to a file `id3c_metadata_with_county.csv` and
saves to the output batch directory. The file is loaded to a dataframe and checked
to make sure all records with source `SFS` have a collection date, with an option
to either continue or bail out if any are missing.