Metadata QC check for matching sample IDs and barcodes
The Excel file containing external metadata contains two sheets which should have
matching lists of sample IDs and barcodes. Adding a QC check to confirm before proceeding.
If there is a mismatch between the two sheets, it should be corrected in the source file
before re-running the `prep_files` script.