Updates to support MIP assembly pipeline
Transition to Molecular Loop assembly pipeline results in different set of input
files for sequencing submissions. A new utils module is being added for functions common
to both the previous and new pipeline. A new file prep script `prep_files_mip.py` was added
to account for the new workflow and lack of metrics and metadata files that had been included
previously.

The `export_id3c_metadata` script was updated to include an option to ignore the `source_origin`
column from the input CSV, rather than use it to filter for retrospectives. With the updated submission
pipeline, the input CSV is filtered prior to calling this script, based on the sample origins provided
by LIMS.