Additional changes related to new MIPS assembly pipeline
As part of the transition to the MIPS assembly pipeline, a number of
values that were previously hard-coded in output files are being parameterized
to handle MIPS or CovidSeq assembly pipelines. The create_submissions
script will require a new argument `--test-name` to indicate whether
to use values for the MIPS or CovidSeq pipeline in these outputs.