Add script to replace IDs in VADR annotation with sequence IDs
In the current submission pipeline, VADR runs on the original FASTA file from assembly, containing identifiers
that concatenate the NWGC ID with the name of the reference genome. For submissions to GenBank, the annotation table
needs to match with the sequence IDs used in the submission FASTA file. This script will replace the appropriate
identifiers and save the annotation table to a new file.